requirements.txt
cryptography>=39.0.0
customtkinter>=5.1.3
pillow>=9.5.0
ttkbootstrap>=1.10.1
bcrypt>=4.0.1
pyperclip>=1.8.2 